Category query
+ Upgrade category admin panel
+ Fix save image bug
+ Привязка рандомных товаров к категориям(для того чтобы картинку-превьюшку показать)
+ Ручная установка картинок категорий
+ Багфикс с полом у категорий
+ Замена enum(прошлая библиотека тупила с сериализацией)

+++ api/models/another_image.py
@@ -34,4 +34,5 @@ def deletePreviousImage(sender, instance, **kwargs):
         except sender.DoesNotExist:
             return
         if old_instance.image != instance.image:
-            default_storage.delete(old_instance.image.path)
+            if old_instance.categoryPreviewImage:
+                default_storage.delete(old_instance.image.path)
